refactor: replace useResidentsNew with useResidents and remove unused files
