+++ src/app/state/rest/index.ts
@@ -36,10 +36,8 @@ export * from "./permissions"
 export * from "./processes"
 export * from "./reasons"
 export * from "./residents"
-export * from "./residentsNew"
 export * from "./roles"
 export * from "./schedules"
 export * from "./tasks"
 export * from "./themes"
-export * from "./tokenOboService"
 export * from "./users"
